requirements.txt
streamlit==1.36.0
pandas==2.2.2
openpyxl==3.1.2
matplotlib==3.8.4
seaborn==0.13.2
python-docx==1.1.0
nltk==3.8.1